refactor(LottoResults): 실수 계산의 정확성을 위해 double -> BigDecimal 전환하기

+++ src/test/java/lotto/domain/LottoGameResultsTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -17,9 +18,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.ONE);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.ONE.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.ONE.getWinningPrize());
     }
 
     @DisplayName("2개가 일치하면 상금은 0원 이다.")
@@ -28,9 +29,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.TWO);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.TWO.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.TWO.getWinningPrize());
     }
 
     @DisplayName("3개가 일치하면 상금은 5000원 이다.")
@@ -39,9 +40,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.THREE);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.THREE.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.THREE.getWinningPrize());
     }
 
     @DisplayName("3개 일치 티켓이 2장이면 상금은 5000원 이다.")
@@ -50,9 +51,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.THREE, LottoRank.THREE);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.THREE.getWinningPrize() * 2);
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.THREE.getWinningPrize() * 2);
     }
 
     @DisplayName("4개가 일치하면 상금은 50000원 이다.")
@@ -61,9 +62,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.FOUR);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.FOUR.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.FOUR.getWinningPrize());
     }
 
     @DisplayName("5개가 일치하면 상금은 1_500_000원 이다.")
@@ -72,9 +73,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.FIVE);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.FIVE.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.FIVE.getWinningPrize());
     }
 
     @DisplayName("5개가 일치하고 보너스번호도 일치하면 상금은 300_000_000원 이다.")
@@ -83,9 +84,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.FIVE_BONUS);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.FIVE_BONUS.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.FIVE_BONUS.getWinningPrize());
     }
 
     @DisplayName("6개가 일치하면 상금은 2_000_000_000원 이다.")
@@ -94,9 +95,9 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.SIX);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double winningPrizeSum = lottoGameResults.getWinningPrizeSum();
+        BigDecimal winningPrizeSum = lottoGameResults.getWinningPrizeSum();
 
-        assertThat(winningPrizeSum).isEqualTo(LottoRank.SIX.getWinningPrize());
+        assertThat(winningPrizeSum.intValue()).isEqualTo(LottoRank.SIX.getWinningPrize());
     }
 
     @DisplayName("결과중 각 당첨 랭크별 개수를 구할 수 있다")
@@ -132,10 +133,10 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.THREE);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double profitRate = lottoGameResults.getProfitRate();
+        BigDecimal profitRate = lottoGameResults.getProfitRate();
         int expected = LottoRank.THREE.getWinningPrize() / PRICE_PER_GAME;
 
-        assertThat(profitRate).isEqualTo(expected);
+        assertThat(profitRate.intValue()).isEqualTo(expected);
     }
 
     @DisplayName("10게임(10000원) 샀을 때 2등 당첨(5개 일치 + 보너스 번호 일치) 1개 로또 수익률 계산 테스트")
@@ -144,10 +145,10 @@ class LottoGameResultsTest {
         List<LottoRank> lottoRanks = Arrays.asList(LottoRank.FIVE_BONUS);
         LottoGameResults lottoGameResults = new LottoGameResults(lottoRanks);
 
-        double profitRate = lottoGameResults.getProfitRate();
+        BigDecimal profitRate = lottoGameResults.getProfitRate();
         int expected = LottoRank.FIVE_BONUS.getWinningPrize() / (lottoRanks.size() * PRICE_PER_GAME);
 
-        assertThat(profitRate).isEqualTo(expected);
+        assertThat(profitRate.intValue()).isEqualTo(expected);
     }
 
 }
